Loan_approval data set
The goal of this project is to develop and evaluate a machine learning model to predict the loan status of people based on their financial records and key information related to loan applications. By analyzing and comparing the model, the aim is to identify the prediction whether a loan will be approved or not.